Supervised_Learning/SVM/SVM_number_predict/requirements.txt
streamlit
scikit-learn
numpy
pillow
joblib

